minio/downloader/requirements.txt
minio
psycopg2-binary